Break up quill_formatted_text into two files.
